feat: replace payment section and action buttons with a unified payment dialog
This introduces a new PaymentDialog component that consolidates payment handling, including method selection and amount input. The previous PaymentSection and ActionButtons components have been removed to streamline the cashier interface.

- Implemented payment dialog with quick cash options
- Added payment method selection and amount input
- Enhanced payment validation and processing feedback
